+++ sympy/polys/dispersion.py
